+++ domain/src/main/java/com/papps/freddy_lazo/domain/model/PetLoverAppointment.java
@@ -14,12 +14,12 @@ public class PetLoverAppointment {
     private String reason;
     private String status;
     private String status_reason;
-    private double qualification;
+    private Integer qualification;
     private String diagnosis;
     private Doctor doctor;
     private List<AppointmentPhoto> photos;
 
-    public PetLoverAppointment(int id, String doctor_id, String pet_lover_id, String pet_by_pet_lover_id, String date, String time, String type, String reason, String status, String status_reason, double qualification, String diagnosis, Doctor doctor,List<AppointmentPhoto> photos) {
+    public PetLoverAppointment(int id, String doctor_id, String pet_lover_id, String pet_by_pet_lover_id, String date, String time, String type, String reason, String status, String status_reason, Integer qualification, String diagnosis, Doctor doctor,List<AppointmentPhoto> photos) {
         this.id = id;
         this.doctor_id = doctor_id;
         this.pet_lover_id = pet_lover_id;
@@ -76,7 +76,7 @@ public class PetLoverAppointment {
         return status_reason;
     }
 
-    public double getQualification() {
+    public Integer getQualification() {
         return qualification;
     }
